requirements.txt
aioredis==1.2.0
amqp==2.5.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
awesome-slugify==1.6.5
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
coverage==4.5.2
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-appconf==1.0.3
django-axes==5.0.9
django-braces==1.13.0
django-celery-beat==1.4.0
django-extensions==2.1.0
django-import-export==1.2.0
django-ipware==2.1.0
django-leaflet==0.24.0
django-model-utils==3.2.0
django-reversion==3.0.4
django-timezone-field==3.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.0
GDAL==2.4.1
greenlet==0.4.15
hiredis==1.0.0
humanize==0.5.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
jsonpickle==1.2
kombu==4.6.3
monotonic==1.5
msgpack==0.6.1
odfpy==1.4.0
openpyxl==2.4.9
psycopg2-binary==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pypiwin32==223
python-crontab==2.3.8
python-dateutil==2.8.0
python-monkey-business==1.0.0
pytz==2019.1
pywin32==224
PyYAML==5.1.1
redis==3.2.1
regex==2019.6.8
requests==2.22.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
Twisted==19.2.0
txaio==18.8.1
tzlocal==1.5.1
Unidecode==0.4.21
urllib3==1.25.3
vboxapi==1.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
